+++ src/components/BounceDetailsContainer/Modals/CancelConfirmationModal/index.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import CenterModal from "@sendgrid/ui-components/center-modal";
 import Button from "@sendgrid/ui-components/button";
-import { Column } from "../../../Column";
-import { Row } from "../../../Row";
+import { Column } from "../../../shared/Column";
+import { Row } from "../../../shared/Row";
 import { WriteSelectors } from "../../selectors";
 
 const ConfirmModalBody = ({ handleModalClose, handleCancelConfirmation }) => (
