+++ src/components/BounceDetailsContainer/index.jsx
@@ -3,17 +3,17 @@ import Breadcrumb from "@sendgrid/ui-components/breadcrumb";
 import Button from "@sendgrid/ui-components/button";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
-import { Row } from "../Row";
-import { Column } from "../Column";
-import Header from "../Header";
-import Pagination from "../Pagination";
+import { Row } from "../shared/Row";
+import { Column } from "../shared/Column";
+import Header from "../shared/Header";
+import Pagination from "../shared/Pagination";
 import DetailsContainer, { DetailsContainerEditable } from "./Details";
 import Changelog from "./Changelog";
 import ChangeModal from "./Modals/ChangeModal";
 import ConfirmationModal from "./Modals/ConfirmationModal";
 import CancelConfirmationModal from "./Modals/CancelConfirmationModal";
 import RevertConfirmationModal from "./Modals/RevertConfirmationModal";
-import NetworkAlert from "../Alerts/NetworkAlert";
+import NetworkAlert from "../shared/Alerts/NetworkAlert";
 import "./index.scss";
 import { WriteSelectors } from "./selectors";
