+++ src/components/shared/Alerts/NetworkAlert/index.scss
