+++ src/components/shared/Column/index.scss
