+++ src/components/shared/Filter/ActivityFilter/index.scss
