+++ src/components/shared/Pagination/index.scss
